Move RegtestData class to own module
